Move the mixin and service to the addon directory.
This allows us to import from a namespaced path and paves the way for
redux-store configuration.
