+++ addon/lib/ember-logger-middleware.js
